backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
azure-common==1.1.28
azure-core==1.39.0
azure-search-documents==11.6.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
isodate==0.7.2
jiter==0.14.0
openai==2.32.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
